Removed cookies containing sensitive data
The use of cookies for authentication and authorization was removed in a
previous commit, but now this commit removes the cookies themselves.

Also changed the field type of the password field on the login page from "text"
to "password". Now, the characters don't appear on the screen when the user is
typing.